Adds task filtering
Implements task filters and the relevant GUI.

TaskFilters class contains TaskFilters which each represent a filter.
Selected filters are shown as "chips" in the GUI, which can be removed
by clicking the "x" button.

Model and Logic classes have additional setters and getters for the
list of available and selected task filters.